correction

+++ solutions/Solution_for_model_6/question_5.cpp
@@ -16,7 +16,7 @@ int main(){
             throw(b);
         }
     }
-    catch(int i){
+    catch(float i){
         cout<<"ERROR"<<endl;
     }
